bot-consumer/requirements.txt
confluent-kafka
python-dotenv
requests
